Move 'product-overview-v2' to 'product-overview' folder and update references to follow new structure.
