+++ client/a8c-for-agencies/sections/marketplace/products-overview/hooks/use-assign-licenses-to-site.ts
