+++ client/a8c-for-agencies/sections/marketplace/products-overview/lib/get-vendor-info.ts
