+++ client/a8c-for-agencies/sections/marketplace/products-overview/product-card/index.tsx
@@ -18,7 +18,7 @@ import withProductLightbox, {
 	WithProductLightboxProps,
 } from '../hocs/with-product-lightbox';
 import ProductBadges from '../product-badges';
-import ProductPriceWithDiscount from '../product-card/product-price-with-discount-info';
+import ProductPriceWithDiscount from './product-price-with-discount-info';
 import useCustomProductCard from './hooks/use-custom-product-card';
 import type { APIProductFamilyProduct } from 'calypso/state/partner-portal/types';
