+++ client/a8c-for-agencies/sections/marketplace/products-overview/product-card/woo-custom-footer/index.tsx
