+++ client/a8c-for-agencies/sections/marketplace/products-overview/product-card/woopayments-revenue-share-notice/style.scss
